// server/prisma/generated/enums.ts
export const PaymentMethod = {
  INSTORE_CASH: 'INSTORE_CASH',
  INSTORE_CARD: 'INSTORE_CARD',
  CASH_APP: 'CASH_APP'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]
// ...
export const TransactionType = {
  DEPOSIT: 'DEPOSIT',
  WITHDRAWAL: 'WITHDRAWAL',
  BET: 'BET',
  WIN: 'WIN',
  TRANSFER_SENT: 'TRANSFER_SENT',
  TRANSFER_RECEIVED: 'TRANSFER_RECEIVED',
  SYSTEM_ADJUSTMENT_CREDIT: 'SYSTEM_ADJUSTMENT_CREDIT',
  SYSTEM_ADJUSTMENT_DEBIT: 'SYSTEM_ADJUSTMENT_DEBIT',
  TOURNAMENT_BUYIN: 'TOURNAMENT_BUYIN',
  TOURNAMENT_PRIZE: 'TOURNAMENT_PRIZE',
  AFFILIATE_COMMISSION: 'AFFILIATE_COMMISSION',
  REFUND: 'REFUND',
  FEE: 'FEE',
  BONUS_AWARD: 'BONUS_AWARD',
  BET_PLACE: 'BET_PLACE',
  BET_WIN: 'BET_WIN',
  BET_LOSE: 'BET_LOSE',
  BET_REFUND: 'BET_REFUND',
  BONUS_WAGER: 'BONUS_WAGER',
  BONUS_CONVERT: 'BONUS_CONVERT',
  BONUS_EXPIRED: 'BONUS_EXPIRED',
  XP_AWARD: 'XP_AWARD',
  ADJUSTMENT_ADD: 'ADJUSTMENT_ADD',
  ADJUSTMENT_SUB: 'ADJUSTMENT_SUB',
  INTERNAL_TRANSFER: 'INTERNAL_TRANSFER',
  PRODUCT_PURCHASE: 'PRODUCT_PURCHASE',
  REBATE_PAYOUT: 'REBATE_PAYOUT'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const TransactionStatus = {
  PENDING: 'PENDING',
  PROCESSING: 'PROCESSING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED',
  CANCELLED: 'CANCELLED',
  REFUNDED: 'REFUNDED',
  EXPIRED: 'EXPIRED',
  REJECTED: 'REJECTED',
  REQUIRES_ACTION: 'REQUIRES_ACTION',
  ON_HOLD: 'ON_HOLD'
} as const

export type TransactionStatus = (typeof TransactionStatus)[keyof typeof TransactionStatus]